+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+## 1.12.0
+### Changed
+ - Control Plane components can now also be discovered using the `tier` and `component` labels, besides `k8s-app`.
+- The integration now uses the infrastructure agent v1.8.23. For more
+  information refer to the [infrastructure agent release notes](https://docs.newrelic.com/docs/release-notes/infrastructure-release-notes/infrastructure-agent-release-notes/)
+  between versions v1.5.75 and v1.8.23.
+
 ## 1.11.0
 ### Changed
  - The old way of determining Leader/Follower status has been switched to a